third_party/blink/web_tests/platform/generic/external/wpt/css/selectors/invalidation/host-pseudo-class-in-has-expected.txt

This is a testharness.js-based test.
PASS Before adding 'a' to #host_parent: Check #subject1 color
PASS Before adding 'a' to #host_parent: Check #subject2 color
FAIL After adding 'a' to #host_parent: Check #subject1 color assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL After adding 'a' to #host_parent: Check #subject2 color assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
PASS After removing 'a' from #host_parent: Check #subject1 color
PASS After removing 'a' from #host_parent: Check #subject2 color
FAIL After adding 'a' to #host: Check #subject1 color assert_equals: expected "rgb(154, 205, 50)" but got "rgb(255, 0, 0)"
FAIL After adding 'a' to #host: Check #subject2 color assert_equals: expected "rgb(135, 206, 235)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
